# app/swagger/swagger.yaml
basePath: /
definitions:
...
  group.DevopsSysGroup:
    properties:
      alias:
        description: 别名
        type: string
      code:
        description: 机构编码
        type: string
      created_at:
        type: string
      domain:
        description: 域
        type: integer
      enable:
        description: 是否启用
        type: integer
      id:
        description: 主键
        type: integer
      linkman:
        description: 联系人
        type: string
      linkmanNo:
        description: 联系人电话
        type: string
      name:
        description: 部门/11111
        type: string
      parentId:
        description: 上级机构
        type: integer
      remark:
        description: 组描述
        type: string
      sort:
        description: 序号
        type: integer
      updated_at:
        type: string
      wechat:
        description: wechat
        type: string
    type: object
...
  operation.DevopsSysOperationRecord:
    properties:
      agent:
        description: 代理
        type: string
      body:
        description: 请求Body
        type: string
      created_at:
        type: string
      error_message:
        description: 错误信息
        type: string
      id:
        description: 主键
        type: integer
      ip:
        description: 请求ip
        type: string
      latency:
        description: 延迟
        type: string
      method:
        description: 请求方法
        type: string
      path:
        description: 请求路径
        type: string
      resp:
        description: 响应Body
        type: string
      status:
        description: 请求状态
        type: integer
      updated_at:
        type: string
      username:
        description: 用户名
        type: string
    type: object
  path.DevopsSysApi:
    properties:
      apiGroup:
        description: api组
        type: string
      createId:
        description: 创建者
        type: integer
      created_at:
        type: string
      description:
        description: api中文描述
        type: string
      id:
        description: 主键
        type: integer
      method:
        description: 方法:创建POST(默认)|查看GET|更新PUT|删除DELETE
        type: string
      path:
        description: api路径
        type: string
      updateId:
        description: 更新人
        type: integer
      updated_at:
        type: string
    type: object
  request.CabinInReceive:
    properties:
      domain:
        type: string
      method:
        type: string
      pType:
        type: string
      resource:
        description: 路径
        type: string
      source:
        type: string
    type: object
...
  request.PageRequest:
    properties:
      filter:
        items:
          type: string
        type: array
      page:
        type: integer
      pageSize:
        type: integer
    type: object
...
  request.SearchApiParams:
    properties:
      apiGroup:
        description: api组
        type: string
      createId:
        description: 创建者
        type: integer
      created_at:
        type: string
      desc:
        description: 排序方式:升序false(默认)|降序true
        type: boolean
      description:
        description: api中文描述
        type: string
      filter:
        items:
          type: string
        type: array
      id:
        description: 主键
        type: integer
      method:
        description: 方法:创建POST(默认)|查看GET|更新PUT|删除DELETE
        type: string
      orderKey:
        description: 排序
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      path:
        description: api路径
        type: string
      updateId:
        description: 更新人
        type: integer
      updated_at:
        type: string
    type: object
...
  request.SearchLogsParams:
    properties:
      agent:
        description: 代理
        type: string
      body:
        description: 请求Body
        type: string
      created_at:
        type: string
      desc:
        description: 排序方式:升序false(默认)|降序true
        type: boolean
      error_message:
        description: 错误信息
        type: string
      filter:
        items:
          type: string
        type: array
      id:
        description: 主键
        type: integer
      ip:
        description: 请求ip
        type: string
      latency:
        description: 延迟
        type: string
      method:
        description: 请求方法
        type: string
      orderKey:
        description: 排序
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      path:
        description: 请求路径
        type: string
      resp:
        description: 响应Body
        type: string
      status:
        description: 请求状态
        type: integer
      timeFilter:
        items:
          type: string
        type: array
      updated_at:
        type: string
      username:
        description: 用户名
        type: string
    type: object
...
  response.PageResult:
    properties:
      columns: {}
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  sys.RelativeRoleApisRequest:
    properties:
      apiIds:
        items:
          type: string
        type: array
      roleId:
        type: string
    type: object
...
  sys.RequestById:
    properties:
      ids:
        description: 角色ID
        type: string
    type: object
...
info:
  contact: {}
  description: This is Devops API
  title: Swagger Devops API
  version: 0.0.1
paths:
  /sys/api/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/path.DevopsSysApi'
      produces:
      - application/json
      responses:
        "200":
          description: 创建基础api
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建基础api
      tags:
      - Apis
  /sys/api/all:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取所有的Api 不分页,返回包括api列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    items:
                      $ref: '#/definitions/path.DevopsSysApi'
                    type: array
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取所有的Api 不分页
      tags:
      - Apis
  /sys/api/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/path.DevopsSysApi'
      produces:
      - application/json
      responses:
        "200":
          description: 删除api
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除api
      tags:
      - Apis
  /sys/api/deletes:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/sys.RequestById'
      produces:
      - application/json
      responses:
        "200":
          description: 删除选中Api
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除选中Api
      tags:
      - Apis
  /sys/api/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: 根据id获取api
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/sys.RequestById'
      produces:
      - application/json
      responses:
        "200":
          description: 根据id获取api,返回包括api详情
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/path.DevopsSysApi'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 根据id获取api
      tags:
      - Apis
  /sys/api/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 分页获取API列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SearchApiParams'
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取API列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取API列表
      tags:
      - Apis
  /sys/api/modify:
    put:
      consumes:
      - application/json
      parameters:
      - description: api路径, api中文描述, api组, 方法
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/path.DevopsSysApi'
      produces:
      - application/json
      responses:
        "200":
          description: 修改基础api
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 修改基础api
      tags:
      - Apis
  /sys/api/role:
    post:
      consumes:
      - application/json
      parameters:
      - description: 给角色关联api 权限
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/sys.RelativeRoleApisRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 给角色关联api权限
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 给角色关联api权限
      tags:
      - Apis
...
  /sys/group/{id}:
    get:
      consumes:
      - application/json
      description: 获得分组接口
      parameters:
      - description: 分组id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获得分组接口
      tags:
      - Group
  /sys/group/add:
    post:
      consumes:
      - application/json
      description: 新增分组接口
      parameters:
      - description: 分组
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/group.DevopsSysGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 新增分组接口
      tags:
      - Group
  /sys/group/add/resources:
    post:
      consumes:
      - application/json
      description: 给分组新增资源接口
      parameters:
      - description: Ptype为p2 , source 是分组的id，resource 是资源， method 为write或者read,或者owner
        in: body
        name: data
        required: true
        schema:
          items:
            $ref: '#/definitions/request.CabinInReceive'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 给分组新增资源接口
      tags:
      - Role
  /sys/group/delete:
    delete:
      consumes:
      - application/json
      description: 删除分组接口
      parameters:
      - description: 分组ids
        in: body
        name: ids
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除分组接口
      tags:
      - Group
  /sys/group/list:
    post:
      consumes:
      - application/json
      description: 获得分组列表接口
      parameters:
      - description: 页数，页大小，筛选条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获得分组列表接口
      tags:
      - Group
  /sys/group/modify:
    post:
      consumes:
      - application/json
      description: 修改分组接口
      parameters:
      - description: 分组
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/group.DevopsSysGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 修改分组接口
      tags:
      - Group
  /sys/group/resource/{name}:
    get:
      consumes:
      - application/json
      description: 获得分组资源接口
      parameters:
      - description: 分组name
        in: path
        name: name
        required: true
        type: string
      - description: 域
        in: query
        name: domain
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获得分组资源接口
      tags:
      - Group
...
  /sys/operation/{id}:
    get:
      consumes:
      - application/json
      description: 获得操作记录接口
      parameters:
      - description: 操作记录的id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获得操作记录接口
      tags:
      - Operation
  /sys/operation/list:
    post:
      consumes:
      - application/json
      description: 获得操作记录列表接口
      parameters:
      - description: 页数，页大小，筛选条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SearchLogsParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/operation.DevopsSysOperationRecord'
            type: array
      security:
      - ApiKeyAuth: []
      summary: 获得操作记录列表接口
      tags:
      - Operation
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: bearer
    type: apiKey
swagger: "2.0"